+++ FlexEngine/include/PlayerController.hpp
@@ -27,7 +27,7 @@ namespace flex
 		BezierCurveList* GetTrackRiding() const;
 
 	private:
-		void SnapPosToTrack();
+		void SnapPosToTrack(real pDistAlongTrack);
 
 		real m_MoveAcceleration = 120.0f;
 		real m_MaxMoveSpeed = 20.0f;
@@ -59,6 +59,7 @@ namespace flex
 		real m_TrackAttachMinDist = 4.0f;
 		// Is true when player began accelerating while facing down the track
 		real m_pDTrackMovement = 0.0f;
+		bool m_bUpdateFacing = false;
 		bool m_bMovingForwardDownTrack = true;
 
 		AudioSourceID m_SoundTrackAttachID;