+++ FlexEngine/include/Track/BezierCurve.hpp
@@ -8,13 +8,12 @@ namespace flex
 		BezierCurve();
 		BezierCurve(const glm::vec3& p0, const glm::vec3& p1, const glm::vec3& p2, const glm::vec3& p3);
 
-		void DrawDebug(bool bHighlighted, const btVector3& highlightColour) const;
+		void DrawDebug(bool bHighlighted, const btVector4& baseColour, const btVector4& highlightColour) const;
 
 		glm::vec3 GetPointOnCurve(real t) const;
-		glm::vec3 GetFirstDerivative(real t) const;
+		glm::vec3 GetCurveDirectionAt(real t) const;
 
-	private:
-		glm::vec3 m_Points[4];
+		glm::vec3 points[4];
 
 	};
 } // namespace flex