+++ changelog.md
@@ -8,3 +8,5 @@ October
 - Prevent input handling when window doesn't have focus
 - Triplanar mapping shader
 - Separate depth-aware editor objects from depth-unaware
+- Add tracks (Bezier Curves which the player can move along for speedy travel)
+- Add track junctions to handle switching tracks